requirements.txt
# Core dependencies for Qwen2.5-VL Weather Forecasting
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.26.0
qwen-vl-utils>=0.0.2

# Image processing
pillow>=10.0.0

# Weather data processing
xarray>=2023.0.0
netcdf4>=1.6.0
dask>=2023.0.0

# Visualization
matplotlib>=3.7.0
cartopy>=0.22.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Optional: For LoRA fine-tuning (Phase 2)
# peft>=0.8.0
# bitsandbytes>=0.41.0
